refactor: design_uiパッケージのcomponents.dartバレルファイルを削除
- components.dartバレルファイルを削除
- src/componentsをlib/componentsに移動
- src/genをlib/genに移動
- 不要なinternal_design_ui_base.dartとinternal_design_ui_test.dartを削除
- 依存ファイルのimport文を更新
