+++ apps/catalog/lib/use_case/internal_design_ui/components/list_tiles.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/widgets.dart';
-import 'package:internal_design_ui/components.dart';
+import 'package:internal_design_ui/components/list_tiles/text_list_tile.dart';
 import 'package:widgetbook/widgetbook.dart';
 import 'package:widgetbook_annotation/widgetbook_annotation.dart';
