+++ packages/design_ui/lib/common_assets.dart
@@ -1,4 +1,4 @@
-import 'package:internal_design_ui/src/gen/assets/assets.gen.dart';
+import 'package:internal_design_ui/gen/assets/assets.gen.dart';
 
 // パッケージ毎に生成されるAssetsクラスと区別するため、直接exportせず別名をつける
 typedef CommonAssets = Assets;
