--- packages/design_ui/lib/components.dart
@@ -1,1 +1,0 @@
-export 'src/components/list_tiles/text_list_tile.dart';
