+++ packages/samples/github_repository/lib/src/ui/github_repository_list_page.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:internal_design_ui/components.dart';
+import 'package:internal_design_ui/components/list_tiles/text_list_tile.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 import 'package:samples_github_repository/src/data/provider/scroll_notifier.dart';
 import 'package:samples_github_repository/src/domain/model/github_repository.dart';
